refactor(st-09056): modularize skills registry
